requirements.txt
# Dependencias del colector de datos de parques

# Base de datos
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Procesamiento de datos
pandas>=2.0.0

# Zona horaria
pytz>=2023.0

# Cliente HTTP (si usas requests en live_client)
requests>=2.31.0

# Opcional: para mejorar rendimiento
# numpy>=1.24.0